Use apache-maintscripts-helper if available
Instead of directly calling a2enconf/a2disconf, use the maintscripts
helper provided by the Apache package if it is present.

This fixes (#68) as it replaces the check for the apache config
directory which can be incorrect if other packages are installed that
include apache config files, such as php-fpm.